% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_candidate_list.R
\name{cluster_candidate_list}
\alias{cluster_candidate_list}
\alias{similar}
\title{Cluster Candidate List}
\usage{
similar(a, b, tol)
}
\arguments{
\item{tol}{integer, tolerance for two numbers being considered similar.}

\item{positions}{vector of candidates in which the main signal might have jumped.}

\item{diff_positions}{derivative of the signal in positions.}
}
\value{
list of candidates grouped in pairs according to clustering and similar jump.
}
\description{
This function clusters candidates in pairs.
}
\examples{
x <- c(rep(1, 10), rep(50, 10), rep(1, 10))
diff_x <- c(0, diff(x))
positions <- which(abs(diff_x) > 10)
}
\keyword{clustering}
