% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_detection.R
\name{diagnose_detection}
\alias{diagnose_detection}
\title{Diagnose Detection}
\arguments{
\item{df}{data.frame containing `x` and `y` coordinates of subject `id`.}
}
\value{
data.frame with id and user input regarding correct
}
\description{
This function is meant to provide a fast plot of the `xy` position trace, `x` and `y` velocities for diagnosis. It can inform the threshold needed for finding wrong detections. User input is required to decide whether the detections are correct (nothing needs to be done) or incorrect (tagged for step detection).
}
\examples{
set.seed(123)
df <- data.frame(id = "first_target", frameID=1:1000, x=rnorm(1000, 0, 5), y = rnorm(1000, 0, 5))
diagnose_detection(df)
}
\keyword{diagnostic}
