% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_step.R
\name{find_step}
\alias{find_step}
\title{Find Step}
\usage{
find_step(x, tolerance, kernel_size = 3)
}
\arguments{
\item{x}{target numeric vector}

\item{tolerance}{integer, threshold distance for step detection after convolution}

\item{kernel_size}{integer, kernel size of the convolution window (by default 3).}
}
\value{
vector of positions of `x` where the step is predicted.
}
\description{
This function tries to use convolution to find steps in data. It returns candidates to remove in a vector.
}
\examples{
x <- c(rep(1,20), rep(50, 20), rep(1, 20))
candidate_10 <- find_step(x, tolerance = 10)
plot(x, type='l')
points(candidate_10, x[candidate_10])
# If tolerance is increased, step is not detected.
candidate_50 <- find_step(x, tolerance = 50)
}
\keyword{convolution}
